WIP: Refactor GitHub actions
The current release workflow uses an unmaintained action. I replaced that with the recommended successor, and also made it automatically pull in the changelog from the latest Umbrel release to save a bit of time when releasing. I'll test this, and we could remove the draft property, so it automatically creates and publishes the release.